+++ changelog.rst
@@ -6,6 +6,7 @@ next release
 
 * Adding missing dev packages on docker images.
 * Docker images clone latest branch of pwnypack repository.
+* Fix magic command registration on IPython >= 5.0.0.
 
 0.8.0 (2015-05-17)
 ==================